% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{aRecord}
\alias{aRecord}
\title{output from an example simulation run}
\format{an array [days,y,x,sex,age]}
\usage{
aRecord
}
\description{
popn by [days,y,x,sex,age] from a simulation run on a simple half grass, half savannah map
}
\keyword{datasets}

