\name{rtAgeing}
\alias{rtAgeing}
\title{ageing of adults or pupae}
\usage{
rtAgeing(v, label)
}
\arguments{
  \item{v}{a vector of age distribution}

  \item{label}{a string describing what is being passed,
  added to any warning messages}
}
\value{
an updated vector
}
\description{
\code{rtAgeing} simply advances the age of all classes in
the passed vector
}

