% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtCalcPupaPerSexAge.r
\name{rtCalcPupaPerSexAge}
\alias{rtCalcPupaPerSexAge}
\title{calculating number of pupae in each age and sex class based on carrying capacity}
\usage{
rtCalcPupaPerSexAge(pMortPupa = 0.25, vpMortF, fStartPopPropCC = 1,
  iCarryCapF = 200)
}
\arguments{
\item{pMortPupa}{pupal mortality per period}

\item{vpMortF}{a vector of age specific F mortality rates, number of age classes is also got from this.}

\item{fStartPopPropCC}{starting population as a proportion of carrying capacity, default = 1}

\item{iCarryCapF}{carrying capacity of adult females}
}
\value{
a vector of abundance by age
}
\description{
\code{rtCalcPupaPerSexAge}
calculates the number of pupae in each and sex class
}

