% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtControlTestGrid1}
\alias{rtControlTestGrid1}
\title{first test of tsetse control}
\usage{
rtControlTestGrid1(aGrid, pControl = 0.1, iControlBorder = 8)
}
\arguments{
\item{aGrid}{array of y,x,sex,age}

\item{pControl}{proportion of adults to be killed by control (daily)}

\item{iControlBorder}{the number of cells around the edge where no control applied
if there are insufficient cells in the passed grid, then it applies control to all cells
(this allows small tests to be run)}
}
\value{
aGrid updated array of y,x,sex,age
}
\description{
\code{rtControlTestGrid1} simply kills off a constant proportion of adults.
In a central control area.
}
\examples{
aRecord <- rtPhase2Test2()
aGrid <- rtGetFromRecord(aRecord,days=1)
aGridControlled <- rtControlTestGrid1(aGrid)
}

