% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtCreateGrid}
\alias{rtCreateGrid}
\title{create a grid array to represent a spatial, sex & age structured popn}
\usage{
rtCreateGrid(nY = 3, nX = 4, sexes = c("F", "M"), nAge = 2, fill = 0)
}
\arguments{
\item{nY}{}

\item{nX}{}

\item{sexes}{= c("F","M")}

\item{nAge}{}

\item{fill}{what value to fill array withs}
}
\value{
a named array with the specified dimensions
}
\description{
\code{rtCreateGrid} creates a named array with dimensions, y,x,sex,age.
The dimensions are named e.g. x1,x2..., y1,y2... to reduce potential for confusion.
}

