% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtCreateRecord}
\alias{rtCreateRecord}
\title{create a record array to store a daily record for a spatial, sex & age structured popn}
\usage{
rtCreateRecord(nDay = 2, nY = 3, nX = 4, sexes = c("F", "M"),
  nAge = 2, fill = 0)
}
\arguments{
\item{nDay}{}

\item{nY}{}

\item{nX}{}

\item{sexes}{= c("F","M")}

\item{nAge}{}

\item{fill}{what value to fill array withs}
}
\value{
a named array with the specified dimensions
}
\description{
\code{rtCreateRecord} creates a named array with dimensions, day,y,x,sex,age.
calls \code{\link{rtCreateGrid}}
The dimensions are named e.g. x1,x2..., y1,y2... to reduce potential for confusion.
}

