% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtDensityDependence}
\alias{rtDensityDependence}
\title{modifies mortality according to density dependence}
\usage{
rtDensityDependence(fPopn, pMort, propDD, iCarryCap)
}
\arguments{
\item{fPopn}{population number, can be adults or pupae}

\item{pMort}{a mortality probability}

\item{propDD}{proportion of mortality that is density dependent}

\item{iCarryCap}{Carrying Capacity as an integer}
}
\value{
pMort modified mortality rate
}
\description{
\code{rtDensityDependence} takes passed mortality rate and modifies it
according to the relationship between density and carrying capacity.
Implemented in the same way as hat-trick.
Allows the proportion of mortality that is density dependent to be specified.
\cr pMort at Carrying Capacity is left unmodified.
\cr pMort at density 0 = pMort*(1-proportion of mortality density dependent).
\cr In between is a linear relationship that continues to give higher values above carrying capacity.
\cr ! I wonder if density dependence could be implemented in a more standard way closer to the logistic model.
}

