% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtFeedingOneHuntPeriod}
\alias{rtFeedingOneHuntPeriod}
\title{tsetse feeding for one cell,age,sex and hunt period}
\usage{
rtFeedingOneHuntPeriod(fHunters = 100, pDetect = 0.1, pFeed = 0.1,
  testing = TRUE)
}
\arguments{
\item{fHunters}{the number of hunting flies in this cell,age,sex and hunt period}

\item{pDetect}{TEMPORARY starting arg, single probability of detection}

\item{pFeed}{TEMPORARY starting arg, single probability of feeding once detected}

\item{testing}{whether to output testing messages to the console}
}
\value{
a list of fManFeeders and ?fHunters
}
\description{
\code{rtFeedingOneHuntPeriod} does feeding for one group of flies in one hunt period.
In the Hat-trick model there were a variable number of hunt periods per day (15-30).
In rtPhase3 the initial proposal is for 15 hunt periods on day 1 & 30 hunt periods on later hunger cycles.
}
\examples{
#if both probs set to 1, returns all feeders
rtFeedingOneHuntPeriod(100,1,1)
#if both probs set to 0, returns all hunters
rtFeedingOneHuntPeriod(100,0,0)
}

