% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtGetFromRecord}
\alias{rtGetFromRecord}
\title{to access population data from a record array [day,x,y,sex,age]}
\usage{
rtGetFromRecord(aRecord, day = "all", x = "all", y = "all", sex = "all",
  age = "all", drop = TRUE)
}
\arguments{
\item{aRecord}{an array with the age distributions of males & females [x,y,sex,age]}

\item{day}{day of the simulation starting at 1}

\item{x}{grid column number}

\item{y}{grid row number}

\item{sex}{'all' returns both sexes separately,'M','F', 'sum' sums sexes}

\item{age}{'all' returns age distribution, 'sum' sums all ages, or an integer age
or an age range too, e.g. c(2:3)}

\item{drop}{whether to drop dimensions that just have a single value, TRUE as default}
}
\value{
an array, matrix or vector named with remaining dimensions of [x,y,sex,age]
}
\description{
\code{rtGetFromRecord} allows access to population data from a grid of sexes and ages.
You can specify which [day,x,y,sex,age] you want to get data for.
Each variable defaults to 'all' so \code{rtGetFromRecord(aRecord)} would return the whole grid.
='sum' can be used to sum across dimensions,
thus \code{rtGetFromRecord(aRecord,x='sum',y='sum',sex='sum',age='sum')}
would produce a single value of the total population on the grid.
}
\examples{
aRecord <- rtPhase2Test3()
aGrid <- rtGetFromRecord(aRecord,day=2) #gives raw array for one day
rtGetFromRecord(aRecord,day=2,x='sum',y='sum',sex='sum') #age structure for whole pop
rtGetFromRecord(aRecord,day=2,x='sum',y='sum',age='sum') #sex ratio for whole pop
#slight anomally this gives 4 grids
rtGetFromRecord(aRecord,day=2,x='all',y='all',age=c(1,2),sex='all')
#this gives just 1
rtGetFromRecord(aRecord,day=2,x='all',y='all',age=c(1,2),sex='sum')
#new test case
aRecord <- rtPhase2Test3(3,3,iDays=4)
rtGetFromRecord(aRecord,day=1:3,x='all',y='all',age=c(1,2),sex='sum')
}

