\name{rtLarvalDeposition}
\alias{rtLarvalDeposition}
\title{tsetse larval deposition}
\usage{
rtLarvalDeposition(vFem, vpDeposit)
}
\arguments{
  \item{vFem}{a vector of the age distribution of Females}

  \item{vpDeposit}{a vector of age-specific deposition
  probabilities of Females}
}
\value{
a list containing vPupF & vPupM
}
\description{
\code{rtLarvalDeposition} returns the number of resulting
pupa male and female
}
\details{
from the age structure of adult females and 'birth'
probabilities per female
}

