% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtMortality}
\alias{rtMortality}
\title{tsetse mortality}
\usage{
rtMortality(vFem, vMal, vpMortF, vpMortM, propDD = 0.25,
  returnArray = FALSE, iCarryCap = NA)
}
\arguments{
\item{vFem}{a vector of the age distribution of Females}

\item{vMal}{a vector of the age distribution of Males}

\item{vpMortF}{a vector of age-specific mortality probabilities of Females}

\item{vpMortM}{a vector of age-specific mortality probabilities of Males}

\item{propDD}{proportion of mortality that is density dependent}

\item{returnArray}{= FALSE, to allow testing new way of returning result as an array}

\item{iCarryCap}{Carrying Capacity as an integer}
}
\value{
a list containing vFem & vMal
}
\description{
\code{rtMortality} returns a list of the age distributions of males & females.
\cr It accepts age and sex specific mortality probabilities.
\cr !It should check that the length of the age structure and pMort vectors are the same
\cr It uses the length of the age structure vectors passed to it.
\cr There could be an option to do F only if only F are passed.
}

