% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtMortalityStableSeek}
\alias{rtMortalityStableSeek}
\title{testing seeking mortality that gives popn stability}
\usage{
rtMortalityStableSeek(iMaxAge = 100, fPopAge1 = 100, fInterval = 0.001,
  fMperF = 0.5, pMortLarva = 0.05, pMortPupa = 0.25, plot = TRUE)
}
\arguments{
\item{iMaxAge}{max age of fly allowed in model}

\item{fPopAge1}{starting popn at age1}

\item{fInterval}{the interval between mortalities used to test}

\item{fMperF}{desired numbers of males per female, default=0.5}

\item{pMortLarva}{larval mortality per period}

\item{pMortPupa}{pupal mortality per period

#'}
}
\value{
float mortality probability
}
\description{
\code{rtMortalityStableSeek}
early development
}

