% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtMoveIsland}
\alias{rtMoveIsland}
\title{movement to cells NESW, island model}
\usage{
rtMoveIsland(m, pMove = 0.4)
}
\arguments{
\item{m}{a matrix of cells containing a single number representing one age}

\item{pMove}{proportion of popn that moves out of the cell.}
}
\value{
an updated matrix following movement
}
\description{
\code{rtMoveIsland} moves proportion of popn in each cell to the 4 neighbouring cells.
Movers are divided equally between the 4 cardinal neighbours.
In this initial version popn moving outside of the grid is lost
and nothing arrives in from outside of the grid.
This function works on a single age class, it can be made to work on multiple age classes
by passing an array[x,y,age] to aaply(.margins=3)
Tries to cope with nrow or ncol==1 but I think fails.
}

