% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtMoveReflect.r
\name{rtMoveReflect}
\alias{rtMoveReflect}
\title{movement to cells NESW, reflecting boundaries}
\usage{
rtMoveReflect(m, pMove = 0.4, verbose = FALSE)
}
\arguments{
\item{m}{a matrix of cells containing a single number representing one age}

\item{pMove}{proportion of popn that moves out of the cell.}

\item{verbose}{print what it's doing T/F}
}
\value{
an updated matrix following movement
}
\description{
\code{rtMoveReflect} moves proportion of popn in each cell to the 4 neighbouring cells.
Movers are divided equally between the 4 cardinal neighbours.
Boundaries are reflecting.
This function works on a single age class, it can be made to work on multiple age classes
by passing an array[y,x,age] to aaply(.margins=3)
Doesn't try to cope with nrow or ncol==1.
}

