% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPhase2Test3}
\alias{rtPhase2Test3}
\title{a simple spatial tsetse population simulation, a 3rd test of phase2}
\usage{
rtPhase2Test3(nCol = 10, nRow = 10, pMove = 0.4, iDays = 4,
  iMaxAge = 120, iCarryCap = 200, iStartAdults = 200, iStartAges = 1,
  iStartPupae = "sameAsAdults", pMortF = 0.05, pMortM = 0.05,
  loop = TRUE, pMortPupa = 0.25, iPupaDensThresh = 200, fSlopeDD = 1,
  iPupDurF = 26, iPupDurM = 28, iFirstLarva = 16, iInterLarva = 10,
  pMortLarva = 0.05, report = NULL)
}
\arguments{
\item{nCol}{number grid columns}

\item{nRow}{number grid rows}

\item{pMove}{probability of moving between cells}

\item{iDays}{days to run simulation}

\item{iMaxAge}{max age of fly allowed in model (will warn if flies age past this)}

\item{iCarryCap}{carrying capacity of adults}

\item{iStartAdults}{number of adults to start simulation with}

\item{iStartAges}{spread start adults across the first n ages classes}

\item{iStartPupae}{number of pupae to start simulation with (they get spread across sex&age)
option "sameAsAdults" to set tot pupae same as tot adults.}

\item{pMortF}{adult female mortality per day}

\item{pMortM}{adult male mortality per day}

\item{loop}{TEMPORARY test of loop versus apply approach, only TRUE works}

\item{pMortPupa}{pupal mortality per period}

\item{iPupaDensThresh}{the threshold pupal density above which density dependence acts}

\item{fSlopeDD}{the slope of density dependence, how mortality increases with density}

\item{iPupDurF}{days it takes pupa(F) to develop}

\item{iPupDurM}{days it takes pupa(M) to develop}

\item{iFirstLarva}{Age that female produces first larva}

\item{iInterLarva}{Inter-larval period}

\item{pMortLarva}{larval mortality per period}

\item{report}{filename for a report for this run, if not specified no report is produced}
}
\value{
a multi-dimensional array [day,x,y,sex,ages]
}
\description{
\code{rtPhase2Test3} goes back to more hat-trick like desnity dependence.
runs a simple spatial popn simulation as a test of phase 2
model components. Concentrates on movement parameters and mortality so that
it can be used to test popn spread under different popn growth rates.
! check if it works when nRow & nCol == 1 (the aspatial model)
}
\examples{
\dontrun{
tst <- rtPhase2Test3()
}
}

