% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPhase3Test}
\alias{rtPhase3Test}
\title{a test of feeding (phase3) across multiple hunt periods}
\usage{
rtPhase3Test(iNumHuntPeriods = 15, fHunters = 1000, pDetectMan = 0.001,
  pDetectOxe = 0.005, pFeedMan = 0.1, pFeedOxe = 0.8, fDensityMan = 1,
  fDensityOxe = 10, testing = TRUE, report = NULL)
}
\arguments{
\item{iNumHuntPeriods}{number of hunt periods}

\item{fHunters}{the number of hunting flies in this cell,age,sex and hunt period}

\item{pDetectMan}{probability of detecting one human per km2}

\item{pDetectOxe}{probability of detecting one other host (ox equivalent) per km2}

\item{pFeedMan}{probability of feeding on a human once detected}

\item{pFeedOxe}{probability of feeding on other host (ox equivalent) once detected}

\item{fDensityMan}{density of humans per km2 in tsetse habitat}

\item{fDensityOxe}{density of other hosts (ox equivalents) per km2 in tsetse habitat}

\item{testing}{whether to output testing messages to the console}

\item{report}{filename for a report for this run, if not specified no report is produced}
}
\value{
dataframe of manFeeders and remaining hunters, length the number of hunt periods
}
\description{
\code{rtPhase3Test} tests feeding for a single group of flies in a single hunger cycle.
The user can set the number of hunt periods, in the main model the plan
is to set the number of hunt periods at 15 for age1 and 30 for subsequent hunger cycles.
}
\examples{
\dontrun{
dF <- rtPhase3Test()
}
}

