% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPhase5Test}
\alias{rtPhase5Test}
\title{simple spatial tsetse population simulation, start popn across whole grid}
\usage{
rtPhase5Test(mCarryCapF = matrix(200, 4, 4), pMove = 0.4, iDays = 4,
  iMaxAge = 120, fMperF = 0.5, fStartPopPropCC = 1, pMortF = 0.05,
  pMortM = 0.05, iMortMinAgeStart = 10, iMortMinAgeStop = 50,
  fMortMinProp = 0.2, fMortOldProp = 0.3, propMortAdultDD = 0.25,
  pMortPupa = 0.25, propMortPupaDD = 0.25, iPupDurF = 26, iPupDurM = 28,
  iFirstLarva = 16, iInterLarva = 10, pMortLarva = 0.05,
  propMortLarvaDD = 0.25, report = NULL)
}
\arguments{
\item{mCarryCapF}{a matrix of female carrying capacities}

\item{pMove}{probability of moving between cells}

\item{iDays}{days to run simulation}

\item{iMaxAge}{max age of fly allowed in model (will warn if flies age past this)}

\item{fMperF}{numbers of males per female, default 0.5 for half as many M as F}

\item{fStartPopPropCC}{starting population as a proportion of carrying capacity, default = 1
option "sameAsAdults" to set tot pupae same as tot adults.}

\item{pMortF}{adult female mortality on day1, rates on later days are determined by following parameters.}

\item{pMortM}{adult male mortality on day1, rates on later days are determined by following parameters.}

\item{iMortMinAgeStart}{Age at which min death rates start.}

\item{iMortMinAgeStop}{Age at which min death rates stop.}

\item{fMortMinProp}{What proportion of the maximum death rate on day 0 is the minimum death rate.}

\item{fMortOldProp}{What proportion of the maximum death rate on day 0 is the death rate after iDeathMinAgeStop.}

\item{propMortAdultDD}{proportion of adult mortality that is density dependent}

\item{pMortPupa}{pupal mortality per period}

\item{propMortPupaDD}{proportion of pupal mortality that is density dependent}

\item{iPupDurF}{days it takes pupa(F) to develop}

\item{iPupDurM}{days it takes pupa(M) to develop}

\item{iFirstLarva}{Age that female produces first larva}

\item{iInterLarva}{Inter-larval period}

\item{pMortLarva}{larval mortality per period}

\item{propMortLarvaDD}{proportion of larval mortality that is density dependent}

\item{report}{filename for a report for this run, if not specified no report is produced}
}
\value{
a multi-dimensional array [day,x,y,sex,ages]
}
\description{
\code{rtPhase5Test} developed from \code{rtPhase2Test3}
Difference being this accepts a grid of carrying capacities,
and initialises popns in all cells on the grid.
I might want to allow nCol,nRow & iCarryCapF to be used instead of mCarryCapF
if mCarryCapF is not passed, then carryCap would be constant across the grid.
}
\examples{
\dontrun{
tst <- rtPhase5Test()
rtPlotMapPop(tst)
}
}

