% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtPhase5Test2}
\alias{rtPhase5Test2}
\title{spatial tsetse population simulation with mortality varying over the grid}
\usage{
rtPhase5Test2(mVegetation = array(c("D", "T", "O", "S", "N", "N"), dim = c(2,
  3)), dfMortByVeg = data.frame(code = c("D", "T", "O", "S", "B", "G", "N"),
  mortality = c(200, 150, 110, 100, 110, 210, 999), pupmortality = c(120, 110,
  105, 100, 120, 170, 999), stringsAsFactors = FALSE), pMove = 0.4,
  iDays = 4, iMaxAge = 120, iCarryCapF = 200, fMperF = 0.5,
  fStartPopPropCC = 1, pMortF = 0.05, pMortM = 0.05,
  iMortMinAgeStart = 10, iMortMinAgeStop = 50, fMortMinProp = 0.2,
  fMortOldProp = 0.3, propMortAdultDD = 0.25, pMortPupa = 0.25,
  propMortPupaDD = 0.25, iPupDurF = 26, iPupDurM = 28, iFirstLarva = 16,
  iInterLarva = 10, pMortLarva = 0.05, propMortLarvaDD = 0.25,
  pControl = 0, verbose = FALSE, report = NULL)
}
\arguments{
\item{mVegetation}{a matrix or filepath for a map of vegetation codes}

\item{dfMortByVeg}{a dataframe or filepath to a lookup table specifying mortality multiplier (percent) for each vegetation type}

\item{pMove}{probability of moving between cells}

\item{iDays}{days to run simulation}

\item{iMaxAge}{max age of fly allowed in model (will warn if flies age past this)}

\item{iCarryCapF}{carrying capacity of adult females}

\item{fMperF}{numbers of males per female, default 0.5 for half as many M as F}

\item{fStartPopPropCC}{starting population as a proportion of carrying capacity, default = 1
option "sameAsAdults" to set tot pupae same as tot adults.}

\item{pMortF}{adult female mortality on day1, rates on later days are determined by following parameters.}

\item{pMortM}{adult male mortality on day1, rates on later days are determined by following parameters.}

\item{iMortMinAgeStart}{Age at which min death rates start.}

\item{iMortMinAgeStop}{Age at which min death rates stop.}

\item{fMortMinProp}{What proportion of the maximum death rate on day 0 is the minimum death rate.}

\item{fMortOldProp}{What proportion of the maximum death rate on day 0 is the death rate after iDeathMinAgeStop.}

\item{propMortAdultDD}{proportion of adult mortality that is density dependent}

\item{pMortPupa}{pupal mortality per period}

\item{propMortPupaDD}{proportion of pupal mortality that is density dependent}

\item{iPupDurF}{days it takes pupa(F) to develop}

\item{iPupDurM}{days it takes pupa(M) to develop}

\item{iFirstLarva}{Age that female produces first larva}

\item{iInterLarva}{Inter-larval period}

\item{pMortLarva}{larval mortality per period}

\item{propMortLarvaDD}{proportion of larval mortality that is density dependent}

\item{pControl}{extra mortality probability due to control, first go at implementing control. TODO improve this}

\item{verbose}{print what it's doing T/F}

\item{report}{filename for a report for this run, if not specified no report is produced}
}
\value{
a multi-dimensional array [day,x,y,sex,ages]
}
\description{
\code{rtPhase5Test2} replaces \code{rtPhase5Test}
}
\details{
This accepts a grid of vegetation and how mortality varies by vegetation
}
\examples{
\dontrun{
tst <- rtPhase5Test2()
rtPlotMapPop(tst)
#now try an unequal matrix
mCarryCapF <- matrix( 10*(1:16), 4, 4)
tst <- rtPhase5Test2(mCarryCapF)
rtPlotMapPop(tst)
#an unequal matrix starting at a fraction of CC
tst <- rtPhase5Test2(mCarryCapF, fStartPopPropCC = 0.5)
rtPlotMapPop(tst)
}
}

