% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtPlotMapVeg.r
\name{rtPlotMapVeg}
\alias{rtPlotMapVeg}
\title{plots a vegetation map (raster)}
\usage{
rtPlotMapVeg(map, title = NULL, colours = c("dark green", "forest green",
  "green1", "light blue", "orange", "pale goldenrod", "white"),
  labels = c("dense forest", "thicket", "open forest", "savannah", "bush",
  "grass", "no go area"), inset = -0.26, pt.cex = 1.7, cex = 0.9,
  bty = "n")
}
\arguments{
\item{map}{either a file or an rtsetse matrix [y,x]}

\item{title}{a title for the plot}

\item{colours}{to use for the vegetation categories, default value gives similar to hat-trick}

\item{labels}{text labels for vegetation categories, default value gives similar to hat-trick}

\item{inset}{for the legend, make it less -ve to move from edge}

\item{pt.cex}{size of boxes in legend}

\item{cex}{size of text in legend}

\item{bty}{whether to have box around legend "o" for yes}
}
\value{
a raster object (numeric) of the passed map
}
\description{
\code{rtPlotMapVeg} plots vegetation map
either from a file or a matrix object
initially accepts hat-trick format vegetation maps
which have character values of : D T O S B G N for 7 categories
}

