\name{rtPlotMeanAge}
\alias{rtPlotMeanAge}
\title{plots mean age of a popn over time}
\usage{
rtPlotMeanAge(df, title = NULL)
}
\arguments{
  \item{df}{a dataframe of number of individuals with days
  in columns and ages in rows}

  \item{title}{a title for the plot}
}
\value{
nothing
}
\description{
\code{rtPlotMeanAge} plots change in mean age from a single
passed dataframe
}
\examples{
tst <- rtPhase1Test(iDays=100)
}

