% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPlotMortRatesByAge}
\alias{rtPlotMortRatesByAge}
\title{plotting age specific mortality rates for Tsetse}
\usage{
rtPlotMortRatesByAge(vpMort, title = "", col = "red")
}
\arguments{
\item{vpMort}{a vector of age specific mortality rates}

\item{title}{a title for the plot}
}
\value{
? nothing
}
\description{
\code{rtPlotMortRatesByAge}
plots mortality rates by Age.
Needs to be called separately for males & females.
}
\examples{
vpMorts <- rtSetMortRatesByAge(c(1:100))
rtPlotMortRatesByAge(vpMorts,"males")
}

