\name{rtPlotPop}
\alias{rtPlotPop}
\title{plots popn over multiple days
\code{rtPlotPop} plots change in pop frompassed single dataframe
you would have to pass 2 dataframes (for M&F to get total popn)}
\usage{
rtPlotPop(df, title = NULL)
}
\arguments{
  \item{df}{a dataframe with days in columns and ages in
  rows}

  \item{title,}{a title for the plot}
}
\value{
nothing
}
\description{
plots popn over multiple days \code{rtPlotPop} plots change
in pop frompassed single dataframe you would have to pass 2
dataframes (for M&F to get total popn)
}

