\name{rtPlotPopAndPupae}
\alias{rtPlotPopAndPupae}
\title{plots popn and pupae over multiple days}
\usage{
rtPlotPopAndPupae(dfF, dfM, dfPupaF, dfPupaM, title = NULL)
}
\arguments{
  \item{dfF}{a dataframe of Males with days in columns and
  ages in rows}

  \item{dfM}{a dataframe of Females with days in columns
  and ages in rows}

  \item{dfPupaF}{a dataframe of Pupa Males with days in
  columns and ages in rows}

  \item{dfPupaM}{a dataframe of Pupa Females with days in
  columns and ages in rows}

  \item{title}{a title for the plot}
}
\value{
nothing
}
\description{
\code{rtPlotPopAndPupae} plots change in pop from passed
dataframes for males and females and pupae
}

