% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtPlotPopGrid}
\alias{rtPlotPopGrid}
\title{plots popn of M&F over multiple days and grid cells}
\usage{
rtPlotPopGrid(aGrid, title = NULL)
}
\arguments{
\item{aGrid}{array of day,y,x,sex,age}

\item{title}{a title for the plot}
}
\value{
?ggplot2 object
}
\description{
\code{rtPlotPopGrid} plots change in M&F numbers over whole grid
from a passed matrix of day,y,x,sex,age.
Uses a different approach from rtPlotPopAndPupae using reshape2, dplyr & ggplot2.
}
\examples{
tst <- rt_runGridTestSpread()
rtPlotPopGrid(tst)
}

