% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtPupalAgeing}
\alias{rtPupalAgeing}
\title{tsetse pupal ageing}
\usage{
rtPupalAgeing(vPupaF, vPupaM)
}
\arguments{
\item{vPupaF}{a vector of the age distribution of Female pupae}

\item{vPupaM}{a vector of the age distribution of Male pupae}
}
\value{
a list containing updated vPupaF & vPupaM
}
\description{
DEPRECATED probably already replaced with rtAgeing
\code{rtPupalAgeing} simply advances the age of all pupa by 1 day
Pupal Emergence is likely to be dealt with in rtPupalEmergence
}

