% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPupalMortalityGrid}
\alias{rtPupalMortalityGrid}
\title{tsetse pupal mortality on a GRID}
\usage{
rtPupalMortalityGrid(aGridPup, pMort = 0.25, propDD = 0.25,
  mCarryCap = NA)
}
\arguments{
\item{aGridPup}{array with the age distributions of pupal males & females [x,y,sex,age]}

\item{pMort}{a mortality probability}

\item{propDD}{proportion of mortality that is density dependent}

\item{mCarryCap}{matrix [x,y] of carrying capacities}
}
\value{
updated aGridPup
}
\description{
\code{rtPupalMortalityGrid} applies \code{\link{rtPupalMortality}} to a grid.
It does that by simply calling that function for each cell in the grid.
}

