% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtPupalMortalityRogers}
\alias{rtPupalMortalityRogers}
\title{tsetse pupal mortality following Rogers(1990)}
\usage{
rtPupalMortalityRogers(vPupaF, vPupaM, pMort = 0.25, iPupaDensThresh = 200,
  fSlopeDD = 1)
}
\arguments{
\item{vPupaF}{a vector of female pupae by age}

\item{vPupaM}{a vector of male pupae by age}

\item{pMort}{a density independent mortality probability}

\item{iPupaDensThresh}{the threshold above which density dependence acts}

\item{fSlopeDD}{the slope of density dependence, how mortality increases with density}
}
\value{
a list of vectors of pupae by age (vPupaF & vPupaM)
}
\description{
\code{rtPupalMortalityRogers} applies pupal mortality following method of Rogers(1990).
\cr Requires 3 parameters, the density independent mortality, threshold that density dependence starts and it's slope (k).
\cr Developed in discussion with John Hargrove
}

