% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtSetAgeStructure.r
\name{rtSetAgeStructure}
\alias{rtSetAgeStructure}
\title{setting an age structure, the abundance in each age class}
\usage{
rtSetAgeStructure(vpMort = rep(0.1, 10), fPopAge0 = 100)
}
\arguments{
\item{vpMort}{a vector of age specific mortality rates, number of age classes is also got from this.}

\item{fPopAge0}{population at age 0}
}
\value{
a vector of abundance by age
}
\description{
\code{rtSetAgeStructure}
fills an age structure based on population at age1 and age-specific mortality rates.
Needs to be called separately for males & females.
}

