% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtSetGridFromVeg.r
\name{rtSetGridFromVeg}
\alias{rtSetGridFromVeg}
\title{setting a grid of anything from a vegetation grid}
\usage{
rtSetGridFromVeg(mVegetation = array(c("D", "T", "O", "S", "N", "N"), dim =
  c(2, 3)), dfLookup = data.frame(from = c("D", "T", "O", "S", "B", "G", "N"),
  to = c(200, 150, 110, 100, 110, 210, 999), stringsAsFactors = FALSE))
}
\arguments{
\item{mVegetation}{a matrix of vegetation types}

\item{dfLookup}{a lookup table specifying a value for each vegetation type}
}
\value{
a grid of converted values
}
\description{
\code{rtSetGridFromVeg} 
creates a grid of anything.
From a vegetation grid and a lookup table.
if the passed matrix contains a code not present in the table it returns NA.
}
\examples{
mTest <- rtSetGridFromVeg()
}

