% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtSetMortGridFromVeg}
\alias{rtSetMortGridFromVeg}
\title{setting a grid of mortality multipliers from a vegetation grid}
\usage{
rtSetMortGridFromVeg(mVegetation = matrix(c("D", "T", "O", "S", "D", "D"),
  nrow = 2), dfMortByVeg = data.frame(code = c("D", "T", "O", "S"), mortality
  = c(100, 200, 300, 400), stringsAsFactors = FALSE))
}
\arguments{
\item{mVegetation}{a matrix of vegetation types}

\item{dfMortByVeg}{a lookup table specifying mortality multiplier (percent) for each vegetation type}
}
\value{
a grid of mortality multipliers
}
\description{
\code{rtSetMortGridFromVeg}
creates a grid of multipliers to be applied to the standard mortality rates.
From a vegetation grid and a table specifying the multiplier for each vegetation.
if the passed matrix contains a code not present in the table it returns NA.
}
\examples{
mMortMult <- rtSetMortGridFromVeg()
}

