\docType{package}
\name{rtsetse}
\alias{rtsetse}
\alias{rtsetse-package}
\title{simulating tsetse fly populations}
\description{
rtsetse provides functions to create tsetse fly population
simulations
}
\details{
Note that package functions refernced using :: don't need
to go here. Just need to go in DESRIPTION. This determines
what imports go in NAMESPACE.
}

