% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtSetVegDifMoveGrids.r
\name{rtSetVegDifMoveGrids}
\alias{rtSetVegDifMoveGrids}
\title{to return an array of grids used internally to represent movement across vegetation boundaries}
\usage{
rtSetVegDifMoveGrids(mVegCats = array(c("O", "O", "O", "O", "S", "O", "O",
  "O", "O", "O", "O", "O"), dim = c(3, 4)), iBestVeg = 4, verbose = FALSE)
}
\arguments{
\item{mVegCats}{a matrix of vegetation categories}

\item{iBestVeg}{the index of the preferred vegetation for this species}

\item{verbose}{print what it's doing T/F}
}
\value{
an array of movement multiplier grids to N,S,E & W, SN,WE,NS,EW
}
\description{
Used to represent effect of vegetation differences between cells.
Can be used to reduce movement into less preferred vegetation.
Returns an array of movement multiplier grids to N,S,E,W, SN,WE,NS,EW.
This is tricky because of the way it is coded to make it run faster.
}

