% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtSetVegMoveGrids.r
\name{rtSetVegMoveGrids}
\alias{rtSetVegMoveGrids}
\title{to return an array of grids used internally to represent movement dependent on vegetation}
\usage{
rtSetVegMoveGrids(mVegCats = array(c("O", "O", "O", "O", "S", "O", "O", "O",
  "O", "O", "O", "O"), dim = c(3, 4)), dfMoveByVeg = data.frame(code = c("D",
  "T", "O", "S", "B", "G", "N"), move = c(0.85, 0.9, 0.95, 1, 1.05, 1.1, 0)),
  mVegMove = NULL, verbose = FALSE)
}
\arguments{
\item{mVegCats}{a matrix of vegetation categories}

\item{dfMoveByVeg}{dataframe specifying a movement multiplier for each vegetation category}

\item{mVegMove}{optional matrix of vegetation movement modifiers >1 increases movement out of the cell, <1 decreases movement out of the cell}

\item{verbose}{print what it's doing T/F}
}
\value{
an array of movement multiplier grids 'here' and to N,S,E & W
}
\description{
Used to represent effect of vegetation in a cell on movement out of that cell.
Returns an array of movement multiplier grids to N,S,E & W
}
\seealso{
\code{\link{rtSetVegDifGrids}} which sets up similar grids for influencing movement according to vegetation differences between cells.\cr
\code{\link{rtMove}} uses the grids created.\cr\cr
The movement vignette contains more details about how movement can be represented, type this in the R console :
\code{vignette("vignette-movement", package="rtsetse")}
}

