% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtTestMortOne}
\alias{rtTestMortOne}
\title{testing seeking mortality that generates stability}
\usage{
rtTestMortOne(fMort, iMaxAge = 100, fPopAge1 = 100)
}
\arguments{
\item{fMort}{test age-independent mortality}

\item{iMaxAge}{max age of fly allowed in model}

\item{fPopAge1}{starting popn at age1}
}
\value{
num larvae
}
\description{
\code{rtTestMortOne}
early development
}

