% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtTestMortSeek}
\alias{rtTestMortSeek}
\title{testing seeking mortality that generates stability}
\usage{
rtTestMortSeek(iMaxAge = 100, fPopAge1 = 100, fInterval = 0.001)
}
\arguments{
\item{iMaxAge}{max age of fly allowed in model}

\item{fPopAge1}{starting popn at age1}

\item{fInterval}{the interval between mortalities used to test}
}
\value{
num larvae
}
\description{
\code{rtTestMortSeek}
early development
}

