% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtMoveReflectNoGoVegBoundary.r
\name{shiftGridIslandN}
\alias{shiftGridIslandN}
\title{shift a matrix one cell N
fill empty cells with 0 for island model (or passed value)}
\usage{
shiftGridIslandN(m, fill)
}
\arguments{
\item{m}{matrix}

\item{fill}{value to put in empty cells, default=0 for island model}
}
\value{
shifted matrix
}
\description{
shift a matrix one cell N
fill empty cells with 0 for island model (or passed value)
}

