% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtMove.r
\name{shiftGridIslandW}
\alias{shiftGridIslandW}
\title{shift a matrix one cell W
fill empty cells with 0 for island model (or passed value)}
\usage{
shiftGridIslandW(m, fill)
}
\arguments{
\item{m}{matrix}

\item{fill}{value to put in empty cells, default=0 for island model}
}
\value{
shifted matrix
}
\description{
shift a matrix one cell W
fill empty cells with 0 for island model (or passed value)
}

