% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtMove.r
\name{shiftGridReflectE}
\alias{shiftGridReflectE}
\title{shift a matrix one cell E, copy boundary to represent reflection}
\usage{
shiftGridReflectE(m)
}
\arguments{
\item{m}{matrix}
}
\value{
shifted matrix
}
\description{
shift a matrix one cell E, copy boundary to represent reflection
}

