% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMove.r
\name{shiftGridReflectN}
\alias{shiftGridReflectN}
\title{shift a matrix one cell N, copy boundary to represent reflection}
\usage{
shiftGridReflectN(m)
}
\arguments{
\item{m}{matrix}
}
\value{
shifted matrix
}
\description{
shift a matrix one cell N, copy boundary to represent reflection
}

