% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMove.r
\name{shiftGridReflectS}
\alias{shiftGridReflectS}
\title{shift a matrix one cell S, copy boundary to represent reflection}
\usage{
shiftGridReflectS(m)
}
\arguments{
\item{m}{matrix}
}
\value{
shifted matrix
}
\description{
shift a matrix one cell S, copy boundary to represent reflection
}

