% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtMoveReflectNoGoVegBoundary.r
\name{shiftGridReflectW}
\alias{shiftGridReflectW}
\title{shift a matrix one cell W, copy boundary to represent reflection}
\usage{
shiftGridReflectW(m)
}
\arguments{
\item{m}{matrix}
}
\value{
shifted matrix
}
\description{
shift a matrix one cell W, copy boundary to represent reflection
}

