% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtAgeing.r
\name{rtAgeing}
\alias{rtAgeing}
\title{ageing of adults or pupae in a single population}
\usage{
rtAgeing(v, label)
}
\arguments{
\item{v}{a vector of age distribution}

\item{label}{a string describing what is being passed, added to any warning messages}
}
\value{
an updated vector
}
\description{
\code{rtAgeing} simply advances the age of all classes in the passed vector.
Note this isn't called by \code{\link{rtAgeingGrid}} which is used for grid models.
}

