% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtAgeingGrid}
\alias{rtAgeingGrid}
\title{ageing of adults or pupae in all cells on a passed grid}
\usage{
rtAgeingGrid(a, label)
}
\arguments{
\item{a}{an array of the age distributions in each grid cell [y,x,sex,age]}

\item{label}{a string describing what is being passed, added to any warning messages}
}
\value{
an updated array
}
\description{
\code{rtAgeingGrid} advances the age of all classes in the vectors in each cell of the passed grid
}

