% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtCreateGrid.r
\name{rtCreateGrid}
\alias{rtCreateGrid}
\title{create a grid array to represent a spatial, sex & age structured popn}
\usage{
rtCreateGrid(nY = 3, nX = 4, sexes = c("F", "M"), nAge = 2, fill = 0)
}
\arguments{
\item{nY}{number of y cells}

\item{nX}{number of x cells}

\item{sexes}{vector of sex names default=c("F","M")}

\item{nAge}{number of ages}

\item{fill}{what value to fill array with}
}
\value{
a named array with the specified dimensions
}
\description{
\code{rtCreateGrid} creates a named array with dimensions, y,x,sex,age.
The dimensions are named e.g. x1,x2..., y1,y2... to reduce potential for confusion.
}
\examples{
aGrid <- rtCreateGrid( nY=3, nX=4,nAge=2, fill=0 )
}

