% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtCreateRecord.r
\name{rtCreateRecord}
\alias{rtCreateRecord}
\title{create a record array to store a daily record for a spatial, sex & age structured popn}
\usage{
rtCreateRecord(nDay = 2, nY = 3, nX = 4, sexes = c("F", "M"),
  nAge = 2, fill = 0)
}
\arguments{
\item{nDay}{number of days}

\item{nY}{number of y cells}

\item{nX}{number of x cells}

\item{sexes}{vector of sex names default=c("F","M")}

\item{nAge}{number of ages}

\item{fill}{what value to fill array with}
}
\value{
a named array with the specified dimensions
}
\description{
\code{rtCreateRecord} creates a named array with dimensions, day,y,x,sex,age.
calls \code{\link{rtCreateGrid}}
The dimensions are named e.g. x1,x2..., y1,y2... to reduce potential for confusion.
To fill array with sequential values after can use :
aRecord[] <- c(1:length(aRecord))
}
\examples{
aRecord <- rtCreateRecord( nDay=2, nY=3, nX=4,nAge=2, fill=0 )
}

