% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtFeedingGrid}
\alias{rtFeedingGrid}
\title{tsetse feeding on a grid}
\usage{
rtFeedingGrid(aGrid, pDetectMan = 0.001, pDetectOxe = 0.005,
  pFeedMan = 0.1, pFeedOxe = 0.8, fDensityMan = 1, fDensityOxe = 10,
  testing = TRUE)
}
\arguments{
\item{aGrid}{an array with the age distributions of males & females [x,y,sex,age]}

\item{pDetectMan}{probability of detecting one human per km2}

\item{pDetectOxe}{probability of detecting one other host (ox equivalent) per km2}

\item{pFeedMan}{probability of feeding on a human once detected}

\item{pFeedOxe}{probability of feeding on other host (ox equivalent) once detected}

\item{fDensityMan}{density of humans per km2 in tsetse habitat}

\item{fDensityOxe}{density of other hosts (ox equivalents) per km2 in tsetse habitat}

\item{testing}{whether to output testing messages to the console}
}
\value{
a list of 2 arrays aGridManFeeders and aGridStarved each with the dimensions [x,y,sex,age]
}
\description{
\code{rtFeedingGrid} will work through the gridded tsetse population by grid cell, sex and age
\cr and work out numbers feeding on himans and numbers starving.
\cr !! it will need to accept grids with the distribution of humans and non-humans
}
\examples{
\dontrun{
tst <- rtPhase2Test2(nRow=3,nCol=3,iMaxAge=7)
aGrid <- rtGetFromRecord(tst, day=2)
fed <- rtFeedingGrid(aGrid)
rtGetFromGrid(fed, sex='sum', age='sum')

#I would expect that this should give similar result to aGrid
#when the probabilities are set to 1
fed <- rtFeedingGrid(aGrid, pDetect=1, pFeed=1)
#not necessarily
#AHA! That's OK it's because only those flies of an age that
#they are on the last day of a feeding cycle will feed.
}
}

