% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtFeedingGrid}
\alias{rtFeedingGrid}
\title{tsetse feeding on a grid **in development**}
\usage{
rtFeedingGrid(aGrid, pDetect = 0.1, pFeed = 0.1)
}
\arguments{
\item{aGrid}{an array with the age distributions of males & females [x,y,sex,age]}

\item{pDetect}{TEMPORARY starting arg, single probability of detection}

\item{pFeed}{TEMPORARY starting arg, single probability of feeding once detected}
}
\value{
a list of 2 arrays aGridManFeeders and aGridStarved each with the dimensions [x,y,sex,age]
}
\description{
\code{rtFeedingGrid} will work through the gridded tsetse population by grid cell, sex and age
\cr and work out numbers feeding on himans and numbers starving.
\cr !! it will need to accept grids with the distribution of humans and non-humans
}
\examples{
\dontrun{
tst <- rtPhase2Test2()
aGrid <- rtGetFromRecord(day=2)
fed <- rtFeedingGrid(aGrid)
rtGetFromGrid(fed, sex='sum', age='sum')
}
}

