% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtLarvalDeposition}
\alias{rtLarvalDeposition}
\title{tsetse larval deposition}
\usage{
rtLarvalDeposition(vFem, vpDeposit)
}
\arguments{
\item{vFem}{a vector of the age distribution of Females}

\item{vpDeposit}{a vector of age-specific deposition probabilities of Females}
}
\value{
a list containing integer numbers of larvae: iLarvaeF, iLarvaeM
}
\description{
\code{rtLarvalDeposition} returns the number of resulting pupa male and female
}
\details{
from the age structure of adult females
and 'birth' probabilities per female.
BEWARE that this round the number of larvae down to an integer. WHich I probably don't need to do.

See \code{\link{rtLarvalDepositionGrid}} for the grid version.
}

