% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtLarvalDeposition.r
\name{rtLarvalDeposition}
\alias{rtLarvalDeposition}
\title{tsetse larval deposition}
\usage{
rtLarvalDeposition(vFem, vpDeposit)
}
\arguments{
\item{vFem}{a vector of the age distribution of Females}

\item{vpDeposit}{a vector of age-specific deposition probabilities of Females}
}
\value{
a list containing numbers of larvae: iLarvaeF, iLarvaeM
}
\description{
\code{rtLarvalDeposition} returns the number of resulting pupa male and female
}
\details{
from the age structure of adult females
and 'birth' probabilities per female.  
It used to round the number of larvae down to an integer, now allows fractional larvae.

See \code{\link{rtLarvalDepositionGrid}} for the grid version.
}

