% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtLarvalDepositionGrid}
\alias{rtLarvalDepositionGrid}
\title{tsetse larval deposition on a grid}
\usage{
rtLarvalDepositionGrid(aGrid, aGridPup, vpDeposit, verbose = FALSE)
}
\arguments{
\item{aGrid}{array with the age distributions of males & females [y,x,sex,age]}

\item{aGridPup}{array with the age distributions of pupal males & females [y,x,sex,age]}

\item{vpDeposit}{a vector of age-specific deposition probabilities of Females}

\item{verbose}{whether to output progress messages to console}
}
\value{
updated array of pupae [y,x,sex,age]
}
\description{
\code{rtLarvalDepositionGrid} returns the number of male and female larvae deposited in each grid cell
}
\details{
from the age structure of adult females
and 'birth' probabilities per female.
}

