% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMortalityGrid.r
\name{rtMortalityGrid}
\alias{rtMortalityGrid}
\title{tsetse mortality on a grid **in development**}
\usage{
rtMortalityGrid(aGrid, vpMortF, vpMortM, propDD = 0.25,
  mMortMultGrid = NULL, mCarryCap = NULL, iCarryCap = NULL)
}
\arguments{
\item{aGrid}{an array with the age distributions of males & females [y,x,sex,age]}

\item{vpMortF}{a vector of age-specific mortality probabilities of Females}

\item{vpMortM}{a vector of age-specific mortality probabilities of Males}

\item{propDD}{proportion of mortality that is density dependent}

\item{mMortMultGrid}{an optional grid of mortality multipliers (derived from vegetation)}

\item{mCarryCap}{a matrix of Carrying Capacities for each cell as an integer (an alternative to iCarryCap)}

\item{iCarryCap}{a single integer Carrying Capacities for all cells as an integer (an alternative to mCarryCap)}
}
\value{
an array with the age distributions of males & females [y,x,sex,age]
}
\description{
\code{rtMortalityGrid} returns an array with the age distributions of males & females [y,x,sex,age]
\cr It accepts age and sex specific mortality probabilities.
\cr !It should check that the length of the age structure and pMort vectors are the same
\cr It uses the length of the age structure vectors passed to it.
\cr There could be an option to do F only if only F are passed.
}

