% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtMortalityStableSeek}
\alias{rtMortalityStableSeek}
\title{seeks day1 mortality rates for F&M that give popn stability}
\usage{
rtMortalityStableSeek(iMaxAge = 100, iTargetPopAge0 = 100,
  fInterval = 0.001, fMperF = 0.5, pMortLarva = 0.05,
  iMortMinAgeStartF = 10, iMortMinAgeStopF = 60, fMortMinPropF = 0.2,
  fMortOldPropF = 0.3, iMortMinAgeStartM = 10, iMortMinAgeStopM = 40,
  fMortMinPropM = 0.2, fMortOldPropM = 0.3, propMortAdultDD = 0.25,
  pMortPupa = 0.25, iPupDurF = 26, iPupDurM = 28, iFirstLarva = 16,
  iInterLarva = 10, plot = TRUE, verbose = FALSE)
}
\arguments{
\item{iMaxAge}{max age of fly allowed in model}

\item{iTargetPopAge0}{target pop of both sexes at age 0}

\item{fInterval}{the interval between mortalities used to test}

\item{fMperF}{desired numbers of males per female, default=0.5}

\item{pMortLarva}{larval mortality per period}

\item{iMortMinAgeStartF}{Age at which min death rates start.}

\item{iMortMinAgeStopF}{Age at which min death rates stop.}

\item{fMortMinPropF}{What proportion of the maximum death rate on day 0 is the minimum death rate.}

\item{fMortOldPropF}{What proportion of the maximum death rate on day 0 is the death rate after iDeathMinAgeStop.}

\item{iMortMinAgeStartM}{Age at which min death rates start.}

\item{iMortMinAgeStopM}{Age at which min death rates stop.}

\item{fMortMinPropM}{What proportion of the maximum death rate on day 0 is the minimum death rate.}

\item{fMortOldPropM}{What proportion of the maximum death rate on day 0 is the death rate after iDeathMinAgeStop.}

\item{propMortAdultDD}{proportion of adult mortality that is density dependent}

\item{pMortPupa}{pupal mortality per period}

\item{iPupDurF}{days it takes pupa(F) to develop}

\item{iPupDurM}{days it takes pupa(M) to develop}

\item{iFirstLarva}{Age that female produces first larva}

\item{iInterLarva}{Inter-larval period}

\item{plot}{whether to output plots}

\item{verbose}{whether to output progress text

#'}
}
\value{
float mortality probability
}
\description{
\code{rtMortalityStableSeek} seeks a age1 mortality rate for F that leads to population stability
and for M that leads to the user supplied ratio of males to females.
It does this by testing successive mortality rates, starting at very low values and increasing until
target values are met.
In each case age1 mortality rates are converted into age dependent mortalities using other user inputs.
The resulting mortalities are used to fill an age structure from age1.
For F, user supplied parameters are used to simulate larval production.
The target is to produce the same number of larvae as age1 pupae - thus leading to a stable population.
Once a stable female age structure has been achieved, the function searches in the same way for
the age1 male mortality rate that generates a male age structure to satisfy the to the user supplied
ratio of males to females.
}

