% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPhase2Test3}
\alias{rtPhase2Test3}
\title{a simple spatial tsetse population simulation, a 3rd test of phase2}
\usage{
rtPhase2Test3(nCol = 10, nRow = 10, pMove = 0.4, iDays = 4,
  iMaxAge = 120, iCarryCapF = 200, fMperF = 0.5, iStartAdults = 200,
  iStartAges = 1, iStartPupae = "sameAsAdults", pMortF = 0.05,
  pMortM = 0.05, iMortMinAgeStart = 10, iMortMinAgeStop = 50,
  fMortMinProp = 0.2, fMortOldProp = 0.3, propMortAdultDD = 0.25,
  pMortPupa = 0.25, propMortPupaDD = 0.25, iPupDurF = 26, iPupDurM = 28,
  iFirstLarva = 16, iInterLarva = 10, pMortLarva = 0.05,
  propMortLarvaDD = 0.25, report = NULL)
}
\arguments{
\item{nCol}{number grid columns}

\item{nRow}{number grid rows}

\item{pMove}{probability of moving between cells}

\item{iDays}{days to run simulation}

\item{iMaxAge}{max age of fly allowed in model (will warn if flies age past this)}

\item{iCarryCapF}{carrying capacity of adult females}

\item{fMperF}{numbers of males per female, default 0.5 for half as many M as F}

\item{iStartAdults}{number of adults to start simulation with}

\item{iStartAges}{spread start adults across the first n ages classes}

\item{iStartPupae}{number of pupae to start simulation with (they get spread across sex&age)
option "sameAsAdults" to set tot pupae same as tot adults.}

\item{pMortF}{adult female mortality on day1, rates on later days are determined by following parameters.}

\item{pMortM}{adult male mortality on day1, rates on later days are determined by following parameters.}

\item{iMortMinAgeStart}{Age at which min death rates start.}

\item{iMortMinAgeStop}{Age at which min death rates stop.}

\item{fMortMinProp}{What proportion of the maximum death rate on day 0 is the minimum death rate.}

\item{fMortOldProp}{What proportion of the maximum death rate on day 0 is the death rate after iDeathMinAgeStop.}

\item{propMortAdultDD}{proportion of adult mortality that is density dependent}

\item{pMortPupa}{pupal mortality per period}

\item{propMortPupaDD}{proportion of pupal mortality that is density dependent}

\item{iPupDurF}{days it takes pupa(F) to develop}

\item{iPupDurM}{days it takes pupa(M) to develop}

\item{iFirstLarva}{Age that female produces first larva}

\item{iInterLarva}{Inter-larval period}

\item{pMortLarva}{larval mortality per period}

\item{propMortLarvaDD}{proportion of larval mortality that is density dependent}

\item{report}{filename for a report for this run, if not specified no report is produced}
}
\value{
a multi-dimensional array [day,x,y,sex,ages]
}
\description{
\code{rtPhase2Test3} goes back to more hat-trick like density dependence.
runs a simple spatial popn simulation as a test of phase 2
model components. Concentrates on movement parameters and mortality so that
it can be used to test popn spread under different popn growth rates.
! check if it works when nRow & nCol == 1 (the aspatial model)
}
\examples{
\dontrun{
tst <- rtPhase2Test3()
rtPlotMapPop(tst)
}
}

