% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtPlotAgeStructure.r
\name{rtPlotAgeStructure}
\alias{rtPlotAgeStructure}
\title{plots age structure over multiple days
\code{rtPlotAgeStructure} visualises change in age structure over time using a raster type plot.
Requires the raster package.}
\usage{
rtPlotAgeStructure(df, title = NULL)
}
\arguments{
\item{df}{a dataframe with days in columns and ages in rows
OR an array of [day,y,x,sex,age]}

\item{title}{a title for the plot}
}
\value{
nothing
}
\description{
plots age structure over multiple days
\code{rtPlotAgeStructure} visualises change in age structure over time using a raster type plot.
Requires the raster package.
}
\examples{
tst <- rt_runAspatial(verbose=FALSE)
rtPlotAgeStructure(tst$dfRecordF) 
}

