\name{rtPlotAgeStructure}
\alias{rtPlotAgeStructure}
\title{plots age structure over multiple days
\code{rtPlotAgeStructure} visualises change in age structure over time using a raster type plot
requires the raster package}
\usage{
rtPlotAgeStructure(df, title = NULL)
}
\arguments{
  \item{df}{a dataframe with days in columns and ages in
  rows}

  \item{title,}{a title for the plot}
}
\value{
nothing
}
\description{
plots age structure over multiple days
\code{rtPlotAgeStructure} visualises change in age
structure over time using a raster type plot requires the
raster package
}

