% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtPlotAgeStructure.r
\name{rtPlotAgeStructure}
\alias{rtPlotAgeStructure}
\title{plots age structure over multiple days}
\usage{
rtPlotAgeStructure(df, title = NULL)
}
\arguments{
\item{df}{a dataframe with days in columns and ages in rows
OR an array of [day,y,x,sex,age]}

\item{title}{a title for the plot}
}
\value{
nothing
}
\description{
\code{rtPlotAgeStructure} visualises change in age structure over time using a raster type plot.
\cr Requires the raster package.
}
\examples{
tst <- rt_runAspatial()
rtPlotAgeStructure(tst$dfRecordF)
}

