% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPlotFeedingOneCycle}
\alias{rtPlotFeedingOneCycle}
\title{plots feeding results of one hunger cycle for testing}
\usage{
rtPlotFeedingOneCycle(dF, title = "Feeding Test")
}
\arguments{
\item{dF}{a dataframe of huntPeriod followed by number of flies hunting or feeding}

\item{title}{a title for the plot}
}
\value{
?ggplot2 object
}
\description{
\code{rtPlotFeedingOneCycle} plots feeding results output by \code{\link{rtPhase3Test}}
}
\examples{
tst <- rtPhase3Test()
rtPlotFeedingOneCycle(tst)
}

