% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtPlotMapVeg}
\alias{rtPlotMapVeg}
\title{plots a vegetation map (raster)}
\usage{
rtPlotMapVeg(map, title = NULL, col = c("white", "dark green",
  "forest green", "green1", "light blue", "orange", "pale goldenrod"),
  veg = c("white", "dark green", "forest green", "green1", "light blue",
  "orange", "pale goldenrod"))
}
\arguments{
\item{map}{either a file or an object}

\item{title}{a title for the plot}

\item{col}{colours to use for the vegetation categories, default value gives similar to hat-trick}

\item{veg}{text labels for vegetation categories, default value gives similar to hat-trick}
}
\value{
a raster object (numeric) of the passed map
}
\description{
\code{rtPlotMapVeg} plots vegetation map
either from a file or a matrix object
initially accepts hat-trick format vegetation maps
which have character values of : D T O S B G N for 7 categories
}

