% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtPlotMeanAge.r
\name{rtPlotMeanAge}
\alias{rtPlotMeanAge}
\title{plots mean age of a popn over time}
\usage{
rtPlotMeanAge(dfF, dfM = NA, title = NULL)
}
\arguments{
\item{dfF}{a dataframe of number of females with days in columns and ages in rows}

\item{dfM}{a dataframe of number of males with days in columns and ages in rows}

\item{title}{a title for the plot}
}
\value{
nothing
}
\description{
\code{rtPlotMeanAge} plots change in mean age passed dataframes for males and females. 
Will also work on just a single dataframe.
}
\examples{
tst <- rt_runAspatial(iDays=100)
}

