% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtPlotPop.r
\name{rtPlotPop}
\alias{rtPlotPop}
\title{plots popn over multiple days}
\usage{
rtPlotPop(df, title = NULL)
}
\arguments{
\item{df}{a dataframe with days in columns and ages in rows}

\item{title}{a title for the plot}
}
\value{
nothing
}
\description{
\code{rtPlotPop} plots change in pop frompassed single dataframe
you would have to pass 2 dataframes (for M&F to get total popn)
}

