% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtPupalMortality}
\alias{rtPupalMortality}
\title{tsetse pupal mortality}
\usage{
rtPupalMortality(vPupaF, vPupaM, pMort = 0.25, propDD = 0.25,
  iCarryCapPupa = NA)
}
\arguments{
\item{vPupaF}{a vector of female pupae by age}

\item{vPupaM}{a vector of male pupae by age}

\item{pMort}{a mortality probability}

\item{propDD}{proportion of mortality that is density dependent}

\item{iCarryCapPupa}{Carrying Capacity as an integer}
}
\value{
a list of vectors of pupae by age (vPupaF & vPupaM)
}
\description{
\code{rtPupalMortality} applies pupal mortality once in the pupal period.
\cr Same rate for males and females (as in Hat-Trick).
\cr Initially this was called for each sex,
but because density dependence needs to know about both sexes I changed so that both are done in same function call.
\cr !Later I may need to pass vegetation type to this or another function that modifies vital rates before passing.
}

