% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtPupalMortalityGrid.r
\name{rtPupalMortalityGrid}
\alias{rtPupalMortalityGrid}
\title{tsetse pupal mortality on a GRID}
\usage{
rtPupalMortalityGrid(aGridPup, pMort = 0.25, propDD = 0.25,
  mMortMultGridPup = NULL, mCarryCap = NULL, iCarryCap = NULL)
}
\arguments{
\item{aGridPup}{array with the age distributions of pupal males & females [y,x,sex,age]}

\item{pMort}{a mortality probability}

\item{propDD}{proportion of mortality that is density dependent}

\item{mMortMultGridPup}{an optional grid of mortality multipliers (derived from vegetation)}

\item{mCarryCap}{a matrix of Carrying Capacities for each cell as an integer (an alternative to iCarryCap)}

\item{iCarryCap}{a single integer Carrying Capacities for all cells as an integer (an alternative to mCarryCap)}
}
\value{
updated aGridPup
}
\description{
\code{rtPupalMortalityGrid} applies \code{\link{rtPupalMortality}} to a grid.
It does that by simply calling that function for each cell in the grid.
}

