% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rtReadMapVeg.r
\name{rtReadMapVeg}
\alias{rtReadMapVeg}
\title{reads a vegetation map (raster) from a text or gridascii file}
\usage{
rtReadMapVeg(inFile)
}
\arguments{
\item{inFile}{filepath}
}
\value{
a matrix (character) of the passed map
TODO it might be better to return a raster, then I can use it to store info from the gridascii header
}
\description{
\code{rtReadMapVeg} reads vegetation map
either from a text or gridascii file
initially accepts hat-trick format vegetation maps
which have character values of : D T O S B G N for 7 categories
}
\examples{
#read in example vegetation map
mVegCats <- rtReadMapVeg( system.file("extdata","vegTanzaniaSerengetiTorr1km.txt", package="rtsetse"))
}

