% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtReportPhase2}
\alias{rtReportPhase2}
\title{produces a report for one run NOTE! this is not used by shiny apps they have their own similar code.}
\usage{
rtReportPhase2(aRecord, lNamedArgs, filename = "reportPhase2.html")
}
\arguments{
\item{aRecord}{output from phase2 a multi-dimensional array for adults [day,y,x,sex,ages] *later it may need to include pupae*}

\item{lNamedArgs}{a list of the arguments and their values passed to rtPhase2Test2}

\item{filename}{a name for the report file}
}
\value{
?nothing maybe
}
\description{
\code{rtReportPhase2} creates a pdf of text and graphical output for this run.
}
\examples{
#tst <- rtPhase2Test2()
#rtReportPhase2(tst, lNamedArgs, filename="myoutput.pdf")
}

