% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtSetMortRatesByAge.r
\name{rtSetMortRatesByAge}
\alias{rtSetMortRatesByAge}
\title{setting age specific mortality rates for Tsetse}
\usage{
rtSetMortRatesByAge(iMaxAge, pMortAge1 = 0.14, iMortMinAgeStart = 10,
  iMortMinAgeStop = 50, fMortMinProp = 0.2, fMortOldProp = 0.3)
}
\arguments{
\item{iMaxAge}{maximum age}

\item{pMortAge1}{mortality rate at day 1 (the maximum)}

\item{iMortMinAgeStart}{Age at which min death rates start.}

\item{iMortMinAgeStop}{Age at which min death rates stop.}

\item{fMortMinProp}{What proportion of the maximum death rate on day 0 is the minimum death rate.}

\item{fMortOldProp}{What proportion of the maximum death rate on day 0 is the death rate after iDeathMinAgeStop.}
}
\value{
a vector of mortality probabilities
}
\description{
\code{rtSetMortRatesByAge} 
creates a vector of mortality rates by Age.
Needs to be called separately for males & females.
\cr I have set defaults in between those for M&F from Hat-trick.
}
\examples{
vpMorts <- rtSetMortRatesByAge(iMaxAge = 100)
rtPlotMortRatesByAge(vpMorts,"males") 
}

